* Makefile, init, config xgpu adjustments
