+++ src/Makefile.am
@@ -12,7 +12,9 @@ AM_CPPFLAGS += -I"/home/sonata/src/4bit_xgpu/xGPU/src"
 
 # Convenience variables to group source files
 hpguppi_databuf = hpguppi_databuf.h \
-		  hpguppi_databuf.c 
+		  hpguppi_databuf.c \
+			hpguppi_xgpu_databuf.c \
+			hpguppi_xgpu_databuf.h
 if BLADE_ENABLED
 hpguppi_databuf += \
 		hpguppi_blade_databuf.c \
@@ -44,7 +46,6 @@ hpguppi_threads = hpguppi_ibverbs_pkt_thread.c \
 		  hpguppi_mb128ch_net_thread.c \
 		  hpguppi_atasnap_ibv_voltage_ftp_thread.c \
 		  hpguppi_atasnap_obs_rawdisk_thread.c \
-		  hpguppi_atasnap_pkt_to_xgpu_transpose.c \
 		  hpguppi_atasnap_xgpu_thread.c \
 		  hpguppi_rawdisk_thread.c \
 		  hpguppi_rawdisk_only_thread.c \
