+++ src/hpguppi_atasnap.h
@@ -162,6 +162,23 @@
 #ifndef _HPGUPPI_ATASNAP_H_
 #define _HPGUPPI_ATASNAP_H_
 
+#define ATA_PAYLOAD_TRANSPOSE_FTP 0
+#define ATA_PAYLOAD_TRANSPOSE_TFP 1
+#define ATA_PAYLOAD_TRANSPOSE_TFP_DF4A 2
+
+#define ATA_PAYLOAD_TRANSPOSE ATA_PAYLOAD_TRANSPOSE_FTP
+
+#if ATA_PAYLOAD_TRANSPOSE == ATA_PAYLOAD_TRANSPOSE_FTP
+#define COPY_PACKET_PAYLOAD_FORLOOP COPY_PACKET_DATA_TO_FTP_DATABUF_FORLOOP
+#define COPY_PACKET_PAYLOAD_DIRECT_FORLOOP COPY_PACKET_DATA_TO_FTP_DATABUF_DIRECT_FORLOOP
+#elif ATA_PAYLOAD_TRANSPOSE == ATA_PAYLOAD_TRANSPOSE_TFP
+#define COPY_PACKET_PAYLOAD_FORLOOP COPY_PACKET_DATA_TO_TFP_DATABUF_FORLOOP
+#define COPY_PACKET_PAYLOAD_DIRECT_FORLOOP COPY_PACKET_DATA_TO_TFP_DATABUF_DIRECT_FORLOOP
+#elif ATA_PAYLOAD_TRANSPOSE == ATA_PAYLOAD_TRANSPOSE_TFP_DF4A
+#define COPY_PACKET_PAYLOAD_FORLOOP COPY_PACKET_DATA_TO_TFP_DF4A_DATABUF_FORLOOP
+#define COPY_PACKET_PAYLOAD_DIRECT_FORLOOP COPY_PACKET_DATA_TO_TFP_DF4A_DATABUF_DIRECT_FORLOOP
+#endif
+
 #include <stdlib.h>
 #include <math.h>
 #include <string.h>
@@ -782,7 +799,7 @@ char align_blk0_with_obsstart(uint64_t * blk0_start_pktidx, uint32_t obsstart, u
 
 typedef struct __attribute__ ((__packed__)) {ATASNAP_DEFAULT_SAMPLE_WIDTH_T num[ATASNAP_DEFAULT_PKTNPOL*ATASNAP_DEFAULT_PKTNTIME];} PKT_DCP_T; // sizeof(PKT_DCP_T) == ATASNAP_DEFAULT_PKT_CHAN_BYTE_STRIDE
 
-#define COPY_PACKET_DATA_TO_FTP_DATABUF_FORLOOP_DIRECT_COPY(\
+#define COPY_PACKET_DATA_TO_FTP_DATABUF_DIRECT_FORLOOP(\
         /*PKT_DCP_T*/  dest_feng_pktidx_offset,/*Indexed into [FENG, PKT_SCHAN, PKTIDX, 0, 0]*/\
         /*PKT_DCP_T*/  pkt_payload,\
         /*const uint16_t*/  pkt_nchan,\
@@ -792,5 +809,5 @@ typedef struct __attribute__ ((__packed__)) {ATASNAP_DEFAULT_SAMPLE_WIDTH_T num[
       *(dest_feng_pktidx_offset) = *pkt_payload++; \
       dest_feng_pktidx_offset += channel_stride;\
     }
-// define COPY_PACKET_DATA_TO_FTP_DATABUF_FORLOOP_DIRECT_COPY
+// define COPY_PACKET_DATA_TO_FTP_DATABUF_DIRECT_FORLOOP
 #endif // _HPGUPPI_ATASNAP_H_
