+++ src/hpguppi_databuf.h
@@ -13,20 +13,14 @@
 // but this keeps things 4K (i.e. page) aligned.
 #define ALIGNMENT_SIZE (4096)
 
-#define N_INPUT_BLOCKS 24
+#define N_INPUT_BLOCKS 36
 #define BLOCK_HDR_SIZE  (5*80*512)      // in bytes, from guppi_daq_server
-#define BLOCK_DATA_SIZE (128*1024*1024) // in bytes, from guppi_daq_server
-#define XGPU_BLOCK_DATA_SIZE 2*BLOCK_DATA_SIZE // in bytes, from guppi_daq_server
+#define BLOCK_DATA_SIZE (120*1024*1024) // in bytes, from guppi_daq_server
 
 typedef struct hpguppi_input_block {
   char hdr[BLOCK_HDR_SIZE];
   char data[BLOCK_DATA_SIZE];
 } hpguppi_input_block_t;
-
-typedef struct hpguppi_input_xgpu_block {
-  char hdr[BLOCK_HDR_SIZE];
-  char data[XGPU_BLOCK_DATA_SIZE];
-} hpguppi_input_xgpu_block_t;
 
 // Used to pad after hashpipe_databuf_t to maintain data alignment
 typedef uint8_t hashpipe_databuf_alignment[
@@ -38,12 +32,6 @@ typedef struct hpguppi_input_databuf {
   hashpipe_databuf_alignment padding; // Maintain data alignment
   hpguppi_input_block_t block[N_INPUT_BLOCKS];
 } hpguppi_input_databuf_t;
-
-typedef struct hpguppi_input_xgpu_databuf {
-  hashpipe_databuf_t header;
-  hashpipe_databuf_alignment padding; // Maintain data alignment
-  hpguppi_input_xgpu_block_t block[N_INPUT_BLOCKS];
-} hpguppi_input_xgpu_databuf_t;
 
 /*
  * INPUT BUFFER FUNCTIONS
@@ -132,99 +120,6 @@ static inline char *hpguppi_databuf_header(struct hpguppi_input_databuf *d, int 
 }
 
 static inline char *hpguppi_databuf_data(struct hpguppi_input_databuf *d, int block_id) {
-    if(block_id < 0 || d->header.n_block < block_id) {
-        hashpipe_error(__FUNCTION__,
-            "block_id %s out of range [0, %d)",
-            block_id, d->header.n_block);
-        return NULL;
-    } else {
-        return d->block[block_id].data;
-    }
-}
-
-hashpipe_databuf_t *hpguppi_input_xgpu_databuf_create(int instance_id, int databuf_id);
-
-static inline hpguppi_input_xgpu_databuf_t *hpguppi_input_xgpu_databuf_attach(int instance_id, int databuf_id)
-{
-    return (hpguppi_input_xgpu_databuf_t *)hashpipe_databuf_attach(instance_id, databuf_id);
-}
-
-static inline int hpguppi_input_xgpu_databuf_detach(hpguppi_input_xgpu_databuf_t *d)
-{
-    return hashpipe_databuf_detach((hashpipe_databuf_t *)d);
-}
-
-static inline void hpguppi_input_xgpu_databuf_clear(hpguppi_input_xgpu_databuf_t *d)
-{
-    hashpipe_databuf_clear((hashpipe_databuf_t *)d);
-}
-
-static inline int hpguppi_input_xgpu_databuf_block_status(hpguppi_input_xgpu_databuf_t *d, int block_id)
-{
-    return hashpipe_databuf_block_status((hashpipe_databuf_t *)d, block_id);
-}
-
-static inline int hpguppi_input_xgpu_databuf_total_status(hpguppi_input_xgpu_databuf_t *d)
-{
-    return hashpipe_databuf_total_status((hashpipe_databuf_t *)d);
-}
-
-static inline int hpguppi_input_xgpu_databuf_wait_free_timeout(
-    hpguppi_input_xgpu_databuf_t *d, int block_id, struct timespec *timeout)
-{
-    return hashpipe_databuf_wait_free_timeout((hashpipe_databuf_t *)d,
-        block_id, timeout);
-}
-
-static inline int hpguppi_input_xgpu_databuf_wait_free(hpguppi_input_xgpu_databuf_t *d, int block_id)
-{
-    return hashpipe_databuf_wait_free((hashpipe_databuf_t *)d, block_id);
-}
-
-static inline int hpguppi_input_xgpu_databuf_busywait_free(hpguppi_input_xgpu_databuf_t *d, int block_id)
-{
-    return hashpipe_databuf_busywait_free((hashpipe_databuf_t *)d, block_id);
-}
-
-static inline int hpguppi_input_xgpu_databuf_wait_filled_timeout(
-    hpguppi_input_xgpu_databuf_t *d, int block_id, struct timespec *timeout)
-{
-    return hashpipe_databuf_wait_filled_timeout((hashpipe_databuf_t *)d,
-        block_id, timeout);
-}
-
-static inline int hpguppi_input_xgpu_databuf_wait_filled(hpguppi_input_xgpu_databuf_t *d, int block_id)
-{
-    return hashpipe_databuf_wait_filled((hashpipe_databuf_t *)d, block_id);
-}
-
-static inline int hpguppi_input_xgpu_databuf_busywait_filled(hpguppi_input_xgpu_databuf_t *d, int block_id)
-{
-    return hashpipe_databuf_busywait_filled((hashpipe_databuf_t *)d, block_id);
-}
-
-static inline int hpguppi_input_xgpu_databuf_set_free(hpguppi_input_xgpu_databuf_t *d, int block_id)
-{
-    return hashpipe_databuf_set_free((hashpipe_databuf_t *)d, block_id);
-}
-
-static inline int hpguppi_input_xgpu_databuf_set_filled(hpguppi_input_xgpu_databuf_t *d, int block_id)
-{
-    return hashpipe_databuf_set_filled((hashpipe_databuf_t *)d, block_id);
-}
-
-static inline char *hpguppi_xgpu_databuf_header(struct hpguppi_input_xgpu_databuf *d, int block_id) {
-    if(block_id < 0 || d->header.n_block < block_id) {
-        hashpipe_error(__FUNCTION__,
-            "block_id %s out of range [0, %d)",
-            block_id, d->header.n_block);
-        return NULL;
-    } else {
-        return d->block[block_id].hdr;
-    }
-}
-
-static inline char *hpguppi_xgpu_databuf_data(struct hpguppi_input_xgpu_databuf *d, int block_id) {
     if(block_id < 0 || d->header.n_block < block_id) {
         hashpipe_error(__FUNCTION__,
             "block_id %s out of range [0, %d)",
